+++ blockfrost-client/CHANGELOG.md
@@ -7,6 +7,8 @@
   * `getScriptRedeemers` for `/scripts/${script_hash}/redeemers` endpoint, principled `getScriptRedeemers'`
   * `getAddressUtxosAsset` for `/addresses/${address}/utxos/${asset}`, principled `getAddressUtxosAsset'`
   * `getScriptDatum` for `/scripts/datum/${datum_hash}` endpoint
+  * `getScriptJSON` for `/scripts/${script_hash}/json` endpoint
+  * `getScriptCBOR` for `/scripts/${script_hash}/cbor` endpoint
 
 # Version [0.1.0.0](https://github.com/blockfrost/blockfrost-haskell/compare/initial...0.1.0.0) (2021-09-14)